making progress with search screen ui
