+++ app/src/main/java/com/csci3397/tigertrails/model/Path.java
@@ -1,0 +1,9 @@
+package com.csci3397.tigertrails.model;
+
+public class Path {
+    String pathName;
+    String description;
+    double distance;
+    double minutes;
+    //TODO: finish
+}
