+++ app/src/main/java/com/csci3397/tigertrails/view/MainActivity.java
@@ -1,15 +1,11 @@
-package com.csci3397.tigertrails;
+package com.csci3397.tigertrails.view;
 
-import androidx.annotation.NonNull;
 import androidx.appcompat.app.AppCompatActivity;
-import androidx.navigation.NavController;
-import androidx.navigation.Navigation;
-import androidx.navigation.ui.AppBarConfiguration;
-import androidx.navigation.ui.NavigationUI;
 
 import android.os.Bundle;
 import android.view.MenuItem;
 
+import com.csci3397.tigertrails.R;
 import com.google.android.material.bottomnavigation.BottomNavigationView;
 import com.google.android.material.navigation.NavigationBarView;
